Sun, Nov 9, 2025, 7:52 PM +00:00

+++ FastAPI_tutorial/app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -20,18 +20,24 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(lifespan=lifespan)
 
-app.include_router(fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"])
-app.include_router(fastapi_users.get_register_router(UserRead, UserCreate), prefix="/auth", tags=["auth"])
-app.include_router(fastapi_users.get_reset_password_router(), prefix="/auth", tags=["auth"])
-app.include_router(fastapi_users.get_verify_router(UserRead), prefix="/auth", tags=["auth"])
-app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix="/users", tags=["users"])
+app.include_router(fastapi_users.get_auth_router(
+    auth_backend), prefix="/auth/jwt", tags=["auth"])
+app.include_router(fastapi_users.get_register_router(
+    UserRead, UserCreate), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(),
+                   prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(
+    UserRead), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_users_router(
+    UserRead, UserUpdate), prefix="/users", tags=["users"])
 
 
 @app.post("/upload")
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),
-    session: AsyncSession = Depends(get_async_session)
+    user: User = Depends(current_active_user),
+    session: AsyncSession = Depends(get_async_session),
 ):
 
     temp_file_path = None
@@ -53,6 +59,7 @@ async def upload_file(
         if upload_result.response_metadata.http_status_code == 200:
 
             post = Post(
+                user_id=user.id,
                 caption=caption,
                 url=upload_result.url,
                 file_type="video" if file.content_type.startswith(
@@ -75,7 +82,8 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user),
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
@@ -95,7 +103,10 @@ async def get_feed(
 
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(
+    post_id: str, session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user),
+):
     try:
         post_uuid = uuid.UUID(post_id)
 
@@ -104,6 +115,9 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(status_code=483, detail="You don't have permission to delete this post")
 
         await session.delete(post)
         await session.commit()
